requirements.txt
# Core dependencies
opencv-python
numpy
Pillow
imagehash
pymongo
scenedetect
faiss-cpu
matplotlib==3.10.3
pytesseract
pydub
python_speech_features
scikit-learn
sentence_transformers
streamlit
pyvis
librosa
#pyplot
ffmpeg-python
openai-whisper
# Optional (if using tqdm progress bars or interactive work)
#tqdm
